Move test parser to separate module
