+++ .tool-versions
@@ -1,0 +1,1 @@
+golang 1.24.0
